- 9.0.8
Update Brix version and enhance streamer method

The Brix version in the pom.xml file has been updated to 9.0.8. In addition, an error-checking conditional has been added to the Streamer class to improve its functionality. This check ensures that data is only written to the output stream if the read operation returns a non-negative integer.